pyproject.toml
[tool.poetry]
name = "ai-chatbot-backend"
version = "0.1.0"
description = ""
authors = ["shubham-codecraft <mane71396@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
cffi = "^1.17.1"
argon2-cffi = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"